requirements.txt
ultralytics
fastapi
uvicorn
ultralytics
pillow
opencv-python-headless
python-multipart